refactor: reorganize hooks into a dedicated hooks directory for better structure and maintainability
