+++ shesha-reactjs/src/designer-components/charts/components/bar/index.tsx
@@ -12,7 +12,7 @@ import {
 import React from 'react';
 import { Bar } from 'react-chartjs-2';
 import { useChartDataStateContext } from '../../../../providers/chartData';
-import { useGeneratedTitle, useIsSmallScreen } from '../../hooks';
+import { useGeneratedTitle, useIsSmallScreen } from '../../hooks/hooks';
 import { IChartData, IChartDataProps } from '../../model';
 import { splitTitleIntoLines, createFontConfig } from '../../utils';
